% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{coef.structest_fit}
\alias{coef.structest_fit}
\title{Extract coefficients from a structest_fit object}
\usage{
\method{coef}{structest_fit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"structest_fit"}.}

\item{...}{further arguments (ignored).}
}
\value{
Named numeric vector of parameter estimates.
}
\description{
Extract coefficients from a structest_fit object
}
