% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_test.R
\name{print.structest_equiv}
\alias{print.structest_equiv}
\title{Print method for structest equivalence tests}
\usage{
\method{print}{structest_equiv}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{"structest_equiv"}.}

\item{digits}{number of significant digits for printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for structest equivalence tests
}
