% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{print.structest_fit}
\alias{print.structest_fit}
\title{Print method for structest_fit objects}
\usage{
\method{print}{structest_fit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{"structest_fit"}.}

\item{digits}{number of significant digits for printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for structest_fit objects
}
