\name{structest-package}
\alias{structest-package}
\docType{package}
\title{Testing the Structural Interpretation of a Latent Factor Model}
\description{
  Implements two statistical tests from VanderWeele and Vansteelandt (2022)
  to reject the structural interpretation of a univariate latent factor model.
}
\references{
  VanderWeele, T. J. and Vansteelandt, S. (2022). A statistical test to
  reject the structural interpretation of a latent factor model.
  \emph{Journal of the Royal Statistical Society Series B}, 84, 1063--1089.
}
