% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.structest}
\alias{summary.structest}
\title{Summary method for structest objects}
\usage{
\method{summary}{structest}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{"structest"}.}

\item{digits}{number of significant digits for printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for structest objects
}
