% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{summary.structest_fit}
\alias{summary.structest_fit}
\title{Summary method for structest_fit objects}
\usage{
\method{summary}{structest_fit}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{"structest_fit"}.}

\item{digits}{number of significant digits for printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for structest_fit objects
}
