% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{vcov.structest_fit}
\alias{vcov.structest_fit}
\title{Variance-covariance matrix for structest_fit parameters}
\usage{
\method{vcov}{structest_fit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"structest_fit"}.}

\item{...}{further arguments (ignored).}
}
\value{
The variance-covariance matrix of the parameter estimates.
}
\description{
Variance-covariance matrix for structest_fit parameters
}
