% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.structest}
\alias{print.structest}
\title{Print method for structest objects}
\usage{
\method{print}{structest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{"structest"}.}

\item{digits}{number of significant digits for printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for structest objects
}
